+++ apps/application/convex/_generated/api.d.ts
@@ -16,21 +16,21 @@ import type * as agent_streamAbort from "../agent/streamAbort.js";
 import type * as agent_streaming from "../agent/streaming.js";
 import type * as agent_textNormalization from "../agent/textNormalization.js";
 import type * as agent_threads from "../agent/threads.js";
-import type * as agent_tools_editEscritoTool from "../agent/tools/editEscritoTool.js";
-import type * as agent_tools_getEscritoStatsTool from "../agent/tools/getEscritoStatsTool.js";
+import type * as agent_tools_editEscritoTool from "../agent/tools/escritos/editEscritoTool.js";
+import type * as agent_tools_getEscritoStatsTool from "../agent/tools/escritos/getEscritoStatsTool.js";
 import type * as agent_tools_getEscritoTool from "../agent/tools/getEscritoTool.js";
 import type * as agent_tools_index from "../agent/tools/index.js";
-import type * as agent_tools_insertContentTool from "../agent/tools/insertContentTool.js";
-import type * as agent_tools_legislationFindTool from "../agent/tools/legislationFindTool.js";
-import type * as agent_tools_legislationReadTool from "../agent/tools/legislationReadTool.js";
+import type * as agent_tools_insertContentTool from "../agent/tools/escritos/insertContentTool.js";
+import type * as agent_tools_legislationFindTool from "../agent/tools/legislation/legislationFindTool.js";
+import type * as agent_tools_legislationReadTool from "../agent/tools/legislation/legislationReadTool.js";
 import type * as agent_tools_listCaseDocumentsTool from "../agent/tools/listCaseDocumentsTool.js";
-import type * as agent_tools_planAndTrackTool from "../agent/tools/planAndTrackTool.js";
-import type * as agent_tools_queryDocumentTool from "../agent/tools/queryDocumentTool.js";
-import type * as agent_tools_readDocumentTool from "../agent/tools/readDocumentTool.js";
+import type * as agent_tools_planAndTrackTool from "../agent/tools/planning/planAndTrackTool.js";
+import type * as agent_tools_queryDocumentTool from "../agent/tools/documents/queryDocumentTool.js";
+import type * as agent_tools_readDocumentTool from "../agent/tools/documents/readDocumentTool.js";
 import type * as agent_tools_readEscritoTool from "../agent/tools/readEscritoTool.js";
-import type * as agent_tools_rewriteEscritoSectionTool from "../agent/tools/rewriteEscritoSectionTool.js";
-import type * as agent_tools_searchCaseDocumentsTool from "../agent/tools/searchCaseDocumentsTool.js";
-import type * as agent_tools_searchLegislationTool from "../agent/tools/searchLegislationTool.js";
+import type * as agent_tools_rewriteEscritoSectionTool from "../agent/tools/escritos/rewriteEscritoSectionTool.js";
+import type * as agent_tools_searchCaseDocumentsTool from "../agent/tools/documents/searchCaseDocumentsTool.js";
+import type * as agent_tools_searchLegislationTool from "../agent/tools/legislation/searchLegislationTool.js";
 import type * as agent_tools_utils from "../agent/tools/utils.js";
 import type * as agent_tools_validation from "../agent/tools/validation.js";
 import type * as auth_utils from "../auth_utils.js";
