+++ apps/application/convex/agent/tools/clients/searchClientsTool.ts
