+++ apps/application/convex/agent/tools/documents/readDocumentTool.ts
@@ -1,9 +1,9 @@
 import { createTool, ToolCtx, getThreadMetadata } from "@convex-dev/agent";
-import { components } from "../../_generated/api";
-import { api, internal } from "../../_generated/api";
+import { components } from "../../../_generated/api";
+import { api, internal } from "../../../_generated/api";
 import { z } from "zod";
-import { getUserAndCaseIds, createErrorResponse, validateStringParam, validateNumberParam } from "./utils";
-import { Id } from "../../_generated/dataModel";
+import { getUserAndCaseIds, createErrorResponse, validateStringParam, validateNumberParam } from "../utils";
+import { Id } from "../../../_generated/dataModel";
 
 /**
  * Tool for reading a document progressively, chunk by chunk.
