+++ apps/application/convex/agent/tools/escritos/getEscritoStatsTool.ts
@@ -1,11 +1,11 @@
 import { createTool, ToolCtx } from "@convex-dev/agent";
-import { internal } from "../../_generated/api";
+import { internal } from "../../../_generated/api";
 import { z } from "zod";
-import { prosemirrorSync } from "../../prosemirror";
-import { buildServerSchema } from "../../../../../packages/shared/src/tiptap/schema";
+import { prosemirrorSync } from "../../../prosemirror";
+import { buildServerSchema } from "../../../../../../packages/shared/src/tiptap/schema";
 import { Node } from "@tiptap/pm/model";
-import { getUserAndCaseIds, createErrorResponse, validateStringParam } from "./utils";
-import { Id } from "../../_generated/dataModel";
+import { getUserAndCaseIds, createErrorResponse, validateStringParam } from "../utils";
+import { Id } from "../../../_generated/dataModel";
 
 /**
  * Statistics object containing document metrics
