+++ apps/application/convex/agent/tools/escritos/manageEscritoTool.ts
