+++ apps/application/convex/agent/tools/escritos/readEscritoTool.ts
