+++ apps/application/convex/agent/tools/escritos/rewriteEscritoSectionTool.ts
@@ -1,8 +1,8 @@
 import { createTool, ToolCtx } from "@convex-dev/agent";
-import { api, internal } from "../../_generated/api";
+import { api, internal } from "../../../_generated/api";
 import { z } from "zod";
-import { Id } from "../../_generated/dataModel";
-import { getUserAndCaseIds, createErrorResponse, validateStringParam } from "./utils";
+import { Id } from "../../../_generated/dataModel";
+import { getUserAndCaseIds, createErrorResponse, validateStringParam } from "../utils";
 
 /**
  * Builds a human-readable description of the anchor configuration
