+++ apps/application/convex/agent/tools/templates/searchTemplatesTool.ts
